Move storage to runtime.py
